requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1
docker==7.1.0
pillow==10.4.0
celery==5.6.2

# UI & Admin UI
django-jazzmin==3.0.1
django-admin-black==1.0.19

# Tools
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
requests==2.32.3
pandas==2.2.3
graphviz==0.20.3
astor==0.8.1 

# AI
anthropic==0.34.2

# Deployment
whitenoise==6.7.0
gunicorn==23.0.0

# DB
#psycopg2-binary==2.9.9
#mysqlclient==2.1.1
django-dbbackup==4.2.1
